例外処理を追加

+++ GetNearRankMod/Managers/ExecuteBatch.cs
@@ -7,10 +7,10 @@ namespace GetNearRankMod.Managers
     internal class ExecuteBatch
     {
         public string _fileName { get; set; } = "NearRank";
+        public string _beatSaberPlaylistPath { get; set; } = $".\\Playlists\\NearRank.bplist";
 
         private readonly string _batchPath = $".\\Libs\\GetNearRank-master\\GetNearRankMod.bat";
         private readonly string _firstPlaylistPath = $".\\playlist.bplist";
-        internal readonly string _beatSaberPlaylistPath = $".\\Playlists\\NearRank.bplist";
 
         ExecuteBatch()
         {
@@ -23,21 +23,37 @@ namespace GetNearRankMod.Managers
         public void ExecuteScrapeBatch()
         {
             Process p = new Process();
-            p.StartInfo.FileName = _batchPath;
-            p.StartInfo.UseShellExecute = false;
-            p.StartInfo.UseShellExecute = false;
-            Logger.log.Debug("Start Generating Playlist");
-            p.Start();
-            p.WaitForExit();
-            p.Close();
-
-            if (File.Exists(_beatSaberPlaylistPath))
+            try
             {
-                File.Delete(_beatSaberPlaylistPath);
+                p.StartInfo.FileName = _batchPath;
+                p.StartInfo.UseShellExecute = false;
+                p.StartInfo.UseShellExecute = false;
+                Logger.log.Debug("Start Generating Playlist");
+                p.Start();
+                p.WaitForExit();
+                p.Close();
+                Logger.log.Debug("Finish Generating Playlist");
             }
-            File.Move(_firstPlaylistPath, _beatSaberPlaylistPath);
+            catch(Exception e)
+            {
+                Logger.log.Debug(e);
+                Logger.log.Critical(@"Something wrong with Beat Saber/Libs/GetNearRank-master/....");
+                p.Close();
+            }
 
-            Logger.log.Debug("Finish Generating Playlist");
+
+            if (File.Exists(_firstPlaylistPath))
+            {
+                if (File.Exists(_beatSaberPlaylistPath))
+                {
+                    File.Delete(_beatSaberPlaylistPath);
+                }
+                File.Move(_firstPlaylistPath, _beatSaberPlaylistPath);
+            }
+            else
+            {
+                Logger.log.Critical($@"{_firstPlaylistPath} does not exist");
+            }
         }
     }
 }
