+++ GetNearRankMod/Override/ChangePlaylistTitleAndImage.cs
@@ -15,16 +15,23 @@ namespace GetNearRankMod.Override
 
         public void AdjustPlaylist()
         {
-            Logger.log.Debug("Start adjustint playlist");
-            StreamReader sr = new StreamReader(_executeBatch._beatSaberPlaylistPath);
-            dynamic json = JsonConvert.DeserializeObject(sr.ReadToEnd());
-            sr.Close();
-            json["playlistTitle"] = _executeBatch._fileName;
-            string js = JsonConvert.SerializeObject(json, Formatting.Indented);
-            StreamWriter sw = new StreamWriter(_executeBatch._beatSaberPlaylistPath, false);
-            sw.WriteLine(js);
-            sw.Close();
-            Logger.log.Debug("Finish adjusting playlist");
+            if (File.Exists(_executeBatch._beatSaberPlaylistPath))
+            {
+                Logger.log.Debug("Start adjustint playlist");
+                StreamReader sr = new StreamReader(_executeBatch._beatSaberPlaylistPath);
+                dynamic json = JsonConvert.DeserializeObject(sr.ReadToEnd());
+                sr.Close();
+                json["playlistTitle"] = _executeBatch._fileName;
+                string js = JsonConvert.SerializeObject(json, Formatting.Indented);
+                StreamWriter sw = new StreamWriter(_executeBatch._beatSaberPlaylistPath, false);
+                sw.WriteLine(js);
+                sw.Close();
+                Logger.log.Debug("Finish adjusting playlist");
+            }
+            else
+            {
+                Logger.log.Critical($@"{_executeBatch._beatSaberPlaylistPath} doesn't exist");
+            }
         }
     }
 }
